added new test script for simple repost detection
Run test2.py interactively in the shell,
and all images in the scraper_cache directory
will be processed.

Input image names can then be given,
and it'll attempt to look for similar images.

It also conveniently generates a repost to
see if it can be detected among the swarm
of other posts.

OCR is not performed at the moment.